Injecting HTML with jQuery - first try
We make use of our helper.js file (which is importantly listed before resumeBuilder.js in index.html). In resumeBuilder.js, we created new variables that reference variables in the helper.js file (which contain lines of HTML) and edit them using .replace(). These new variables (which are lines of HTML with content which we modified) are then injected into our HTML using jQuery's .prepend().