Created new functions, an event handler, a button
Created a new displayWork() function that appends all of the formatted work info.

Used a jQuery event handler to store and log user click locations, using the .click() method and a function in helper.js called logClicks(x,y).

Appended a button to the main html div (the button was already created in helper.js).

Created a new inName(names) function that "internationalizes" a given name. This function is linked to the click of the button we appended to the page with the help of some jQuery already in helper.js.